refactor(cms): update lists/pdf.ts. Remove Google Docs PDF viewer use <iframe src="${pdfURL}"> instead.
In older browsers or devices, we are still encountering the issue where
the Google Docs PDF viewer shows a blank page in the iframe. The main
reason seems to be that Google Docs does not always return a 200
response, which would allow us to successfully view the PDF. Sometimes,
Google Docs returns a 204 response, leading to the iframe displaying a
blank page.

We currently suspect that the 204 response from Google Docs might be
related to the frequency of our requests.